+++ packages/api/src/textToSpeech.d.ts
@@ -1,21 +1,29 @@
 declare module '@omnivore/text-to-speech-handler' {
-  export function htmlToSsmlItems(
+  export function htmlToSpeechFile(
     html: string,
     options: SSMLOptions
-  ): SSMLItem[]
+  ): SpeechFile
 
   export interface SSMLOptions {
-    primaryVoice: string
-    secondaryVoice: string
-    rate: string
-    language: string
+    primaryVoice?: string
+    secondaryVoice?: string
+    rate?: number
+    language?: string
   }
 
-  export interface SSMLItem {
-    open: string
-    close: string
-    textItems: string[]
+  interface Utterance {
     idx: number
+    wordOffset: number
+    wordCount: number
     voice?: string
+    text: string
+  }
+
+  export interface SpeechFile {
+    wordCount: number
+    averageWPM: number
+    language: string
+    defaultVoice: string
+    utterances: Utterance[]
   }
 }
